Abstract github interface out of spr script
This makes unit testing spr main logic possible

commit-id:47e99573
