+++ cmd/spr/main.go
@@ -6,12 +6,12 @@ import (
 	"os"
 
 	"github.com/ejoffe/rake"
+	"github.com/ejoffe/spr/config"
+	"github.com/ejoffe/spr/github/githubclient"
 	"github.com/ejoffe/spr/spr"
 	flags "github.com/jessevdk/go-flags"
 	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
-	"github.com/shurcooL/githubv4"
-	"golang.org/x/oauth2"
 )
 
 var (
@@ -49,48 +49,30 @@ func main() {
 		os.Exit(0)
 	}
 
+	//  check that we are inside a git dir
 	err = spr.SanityCheck()
 	if err != nil {
 		fmt.Println(err)
 		os.Exit(2)
 	}
 
-	if opts.Debug {
-		zerolog.SetGlobalLevel(zerolog.DebugLevel)
-	}
-
 	// parse configuration
-	cfg := spr.Config{}
+	cfg := config.Config{}
 	rake.LoadSources(&cfg,
 		rake.DefaultSource(),
-		spr.GitHubRemoteSource(&cfg),
-		rake.YamlFileSource(spr.ConfigFilePath()),
-		rake.YamlFileWriter(spr.ConfigFilePath()),
+		config.GitHubRemoteSource(&cfg),
+		rake.YamlFileSource(config.ConfigFilePath()),
+		rake.YamlFileWriter(config.ConfigFilePath()),
 	)
+
 	if opts.Debug {
+		zerolog.SetGlobalLevel(zerolog.DebugLevel)
 		rake.LoadSources(&cfg, rake.DebugWriter(os.Stdout))
 	}
 
 	ctx := context.Background()
-
-	token := os.Getenv("GITHUB_TOKEN")
-	if token == "" {
-		fmt.Printf("GitHub OAuth Token Required\n")
-		fmt.Printf("Make one at: https://%s/settings/tokens\n", "github.com")
-		fmt.Printf("And set an env variable called GITHUB_TOKEN with it's value\n")
-		os.Exit(3)
-	}
-	ts := oauth2.StaticTokenSource(
-		&oauth2.Token{AccessToken: token},
-	)
-	tc := oauth2.NewClient(ctx, ts)
-
-	client := githubv4.NewClient(tc)
-
+	client := githubclient.NewGitHubClient(ctx, &cfg)
 	stackedpr := spr.NewStackedPR(&cfg, client, os.Stdout, opts.Debug)
-	if opts.Debug {
-		zerolog.SetGlobalLevel(zerolog.DebugLevel)
-	}
 
 	if opts.Update {
 		stackedpr.UpdatePullRequests(ctx)
