+++ github/pullrequest_test.go
@@ -1,0 +1,71 @@
+package github
+
+import (
+	"testing"
+
+	"github.com/ejoffe/spr/config"
+)
+
+func TestSortPullRequests(t *testing.T) {
+	prs := []*PullRequest{
+		{
+			Number:     3,
+			FromBranch: "third",
+			ToBranch:   "second",
+		},
+		{
+			Number:     2,
+			FromBranch: "second",
+			ToBranch:   "first",
+		},
+		{
+			Number:     1,
+			FromBranch: "first",
+			ToBranch:   "master",
+		},
+	}
+
+	config := &config.Config{}
+	prs = SortPullRequests(prs, config)
+	if prs[0].Number != 1 {
+		t.Fatalf("prs not sorted correctly %v\n", prs)
+	}
+	if prs[1].Number != 2 {
+		t.Fatalf("prs not sorted correctly %v\n", prs)
+	}
+	if prs[2].Number != 3 {
+		t.Fatalf("prs not sorted correctly %v\n", prs)
+	}
+}
+
+func TestSortPullRequestsMixed(t *testing.T) {
+	prs := []*PullRequest{
+		{
+			Number:     3,
+			FromBranch: "third",
+			ToBranch:   "second",
+		},
+		{
+			Number:     1,
+			FromBranch: "first",
+			ToBranch:   "master",
+		},
+		{
+			Number:     2,
+			FromBranch: "second",
+			ToBranch:   "first",
+		},
+	}
+
+	config := &config.Config{}
+	prs = SortPullRequests(prs, config)
+	if prs[0].Number != 1 {
+		t.Fatalf("prs not sorted correctly %v\n", prs)
+	}
+	if prs[1].Number != 2 {
+		t.Fatalf("prs not sorted correctly %v\n", prs)
+	}
+	if prs[2].Number != 3 {
+		t.Fatalf("prs not sorted correctly %v\n", prs)
+	}
+}
